NoteRange iterable now
